WIP: Alphanet/DevNet tests fixing
